[SERVER] - follow 기능 완료

+++ server/src/follow/follow.service.ts
@@ -2,7 +2,8 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { Follow } from '@prisma/client';
 import { UserService } from 'src/user/user.service';
-import { FolloweeNotFoundException } from './exceptions/followeeNotFound.exception';
+import { FollowNotFoundException } from './exceptions/followNotFound.exception';
+import { FollowYourselfForbiddenException } from './exceptions/followYourselfForbidden.exception';
 @Injectable()
 export class FollowService {
   constructor(
@@ -12,11 +13,11 @@ export class FollowService {
 
   async followUser(followerId: number, followingId: number): Promise<Follow> {
     if (followerId === followingId) {
-      throw new Error('자기 자신을 팔로우할 수 없습니다.');
+      throw new FollowYourselfForbiddenException();
     }
 
-    if (!(await this.isExistFollowee(followerId, followingId))) {
-      throw new FolloweeNotFoundException();
+    if (await this.isExistFollowee(followerId, followingId)) {
+      throw new FollowNotFoundException();
     }
 
     return await this.prisma.follow.create({
@@ -24,48 +25,36 @@ export class FollowService {
     });
   }
 
-  async unfollowUser(followerId: number, followingId: number) {
-    return await this.prisma.follow.deleteMany({
+  async unfollowUser(
+    followerId: number,
+    followingId: number,
+  ): Promise<boolean> {
+    console.log(followerId, followingId);
+    if (await this.isExistFollowee(followerId, followingId)) {
+      throw new FollowNotFoundException();
+    }
+
+    const result = await this.prisma.follow.deleteMany({
       where: { followerId, followingId },
     });
+
+    return result.count > 0 ? true : false;
   }
 
-  async getFollowers(userId: number) {
+  async getFollowers(userId: number): Promise<Follow[]> {
     const followers = await this.prisma.follow.findMany({
       where: { followingId: userId },
-      include: {
-        follower: {
-          include: {
-            _count: { select: { followers: true, followings: true } },
-          },
-        },
-      },
     });
 
-    return followers.map((follow) => ({
-      user: follow.follower,
-      followerCount: follow.follower._count.followers,
-      followingCount: follow.follower._count.followings,
-    }));
+    return followers;
   }
 
-  async getFollowing(userId: number) {
+  async getFollowing(userId: number): Promise<Follow[]> {
     const following = await this.prisma.follow.findMany({
       where: { followerId: userId },
-      include: {
-        following: {
-          include: {
-            _count: { select: { followers: true, followings: true } },
-          },
-        },
-      },
     });
 
-    return following.map((follow) => ({
-      user: follow.following,
-      followerCount: follow.following._count.followers,
-      followingCount: follow.following._count.followings,
-    }));
+    return following;
   }
 
   async getFollowerCount(userId: number): Promise<number> {
@@ -83,6 +72,6 @@ export class FollowService {
     const followerUser = await this.userService.findOneById(followerId);
     const followingUser = await this.userService.findOneById(followingId);
 
-    return !followerUser && !followingUser;
+    return !followerUser || !followingUser;
   }
 }
